Improve binary search algorithm
If a directory of the same name is in the current directory, this preventing the search running.  So, if the binary path is not fully qualified always search, and check to ensure it is a file that is found, rather than an executable directory.
Thanks to Ian Marsh for the patch