Add 'auto quit' setting to the File menu
The initial value for "auto quit" is taken from the config file but can
be overridden within the console UI for that session.

Although menu code for "auto close" is added, the functionality does not
work as expected.  This is because the setting is put onto each terminal
when it is opened and is not updated thereafter.